requirements.txt
streamlit==1.22.0
pandas==1.5.3
plotly==5.14.1
beautifulsoup4==4.12.2
transformers==4.28.1
nltk==3.8.1
requests==2.29.0
langdetect==1.0.9
numpy==1.24.3
torch==2.0.1
psycopg2-binary==2.9.5
python-dotenv==1.0.0